cfr/requirements.txt
# Core dependencies
torch>=2.6        # CUDA and CPU support included; for Intel Arc/XPU also install intel_extension_for_pytorch
numpy>=2.0
cffi>=2.0
pydantic>=2.0
PyYAML>=6.0
joblib>=1.3
tqdm>=4.65
psutil>=5.9
rich>=13.0
typer[all]>=0.9
